Update README.md. Lint code and documentation.
